refactor client router initialization (#67821)
Previously, the `AppRouter` component was responsible for creating the
initial router state, while the `ActionQueue` was initialized in
`app-index`. This led to a weird inversion of control: `ActionQueue`
could technically have `null` state, which was invalid and guarded by an
invariant. To account for this, the state was lazily initialized on the
first `dispatch`. We also had to have separate exports & `isServer`
conditions since `AppRouter` is also rendered during SSR.

With the refactor in #64594, the render stream no longer includes the
`AppRouter` component. This means that we can initialize the client
router state in `app-index` using `initialRSCPayload`, and create the
`ActionQueue` with that initial state. That way `AppRouter` always has
an initialized `ActionQueue`.

Closes NDX-92