docs: add i18n caveat for global-error file #66905 (#67894)
### What?

Clarify the location of the `global-error.tsx` file in the documentation

Related PR:
https://github.com/vercel/next.js/pull/66908

Related issues:
https://github.com/vercel/next.js/issues/61400
https://github.com/vercel/next.js/issues/60245

### Why?

Many users, including myself, find it unintuitive from the [official
documentation](https://nextjs.org/docs/app/building-your-application/routing/internationalization)
about i18n and the example on
[github.com](https://github.com/vercel/next.js/tree/canary/examples/app-dir-i18n-routing)
that the `global-error.tsx` file should be placed in the root of the
`app/` directory instead of `app/[lang]/`. The documentation and example
**do not cover this detail**, which leads users to place
`global-error.tsx` in `app/[lang]/`. As a result, when
`global-error.tsx` **cannot catch errors** in root `layout.tsx`, users
report this as a new issue on Next.js's GitHub, as seen in the related
issues linked above.

Fixes #66905

Co-authored-by: Sam Ko <sam@vercel.com>