Prevent unnecessary rsc reloads on navigation
In larger apps, when compilation can take time and endpoints are often build concurrently, `endpoint_server_changed_subscribe`’s subscription to turbotasks can result in unnecessary calls of the JS callback even when turbotasks input has not changed.

This wraps subscribing to issues and diagnostics in `endpoint_server_changed_subscribe` within a turbotasks function that ensures that inputs are properly consistent (e.g. diagnostics have been sorted and are value-equal) so this callback is not called unnecessary.

Test Plan: Navigate within a large app and verify `serverComponentChanges` messages are not sent through the web socket connection and the browser does not perform unnecessary fetch requests to rsc endpoints.