Do not limit the number of fetch requests that are logged
Previously, we limited the logged fetch requests to the 11 slowest (10
was intended, off-by-one error). There was no indication given to the
user, though, that the logged requests may not be a complete
representation of all fetch requests made. This lack of transparency
could lead to confusion or misinterpretation of the fetch metrics.

Additionally, we had a deduplication logic for logged fetch requests
based on `url`, `status`, and `method`. This approach inadvertently
caused the logs of POST requests with different request bodies to be
deduped.

To improve clarity and provide a more comprehensive overview, we will
now log all fetch requests without any arbitrary limits. We have also
removed the deduplication logic to ensure that all unique request logs
are preserved. This ensures that users have full visibility into the
fetch activity and can accurately diagnose and troubleshoot any issues
related to fetch performance.

fixes #51699