Add experimental flag for chunk failure handling during app navigation (#67926)
This ensures when we encounter chunk load errors or uncaught
rejections/exceptions during a client transition in app router that we
properly fallback to a hard navigation which allows us to recover
instead of staying in a bad state.

Currently when a chunk fails to load it in most cases hits the error
boundary and that will be rendered (one bad state), another case that
can occur is the chunk fails to load while react-server-dom-webpack is
handling the stream and it crashes and the error boundary isn't even
triggered (second bad state). Both of which aren't recovered until the
user manually navigates.

This is currently behind a flag to allow testing/validating. 

Closes: NDX-75

---------

Co-authored-by: Zack Tanner <1939140+ztanner@users.noreply.github.com>