Avoid doing the work of `createNextInstall` twice
The `run-tests.js` script already calls `createNextInstall` and does all
the heavy lifting. Calling it again in the `create-next-app` integration
tests is wasteful.

In addition, the built `next` tarball is exposed by `run-tests.js` via
the `NEXT_TEST_PKG_PATHS` env variable. We need to provide the filename
of the tgz as `NEXT_PRIVATE_TEST_VERSION` so that `create-next-app` uses
it in the generated `package.json` as the `next` dependency "version".

With the previous approach, the installation was pointing `next` at the
`installDir`, which has it's own version of React installed. This
created runtime errors (`TypeError: Cannot read properties of null
(reading 'useContext')`) when running `next dev` in a pages project that
was created with `create-next-app`.