Ensure client components are in experimental trace (#68492)
This ensures we properly trace client components with the experimental
tracing as currently the trace plugin is failing to detect these
dependencies due to the `flight-client-module` proxy transform taking
place before the trace entrypoints plugin is applied. To avoid this
issue it grabs the original source and only applies the swc loader for
now which handles the majority case. We can investigate in the future
applying other loaders as well to handle more edge cases.

Closes: NDX-135
Closes: NDX-134