[turbopack] Replace uses of globby in scripts with glob (#68493)
Replace uses of `globby` (which is not in the `package.json`) with
`glob` (which is in the `package.json`).

These scripts were part of the nextpack repository (RIP), and
@wbinnssmith copied them over in #68471

Tested by running `pnpm pack-next`, inspecting the `tarballs` directory,
installing the tarballs into shadcn-ui, and building with turbopack.

`globby` and `glob` share the same API, but Next.js pulls in a pretty
old version of `glob`, so some of the newer APIs don't exist, and it
doesn't natively use promises:
https://github.com/isaacs/node-glob/tree/v7.1.6

We could upgrade `glob` without too much pain... It's only used for
`devDependencies`.