only perform VersionedContentMap bookkeeping in dev (#68445)
### What?

We have a data structure VersionedContentMap that is used in dev mode to manage hmr and source maps. This is irrelevant in prod so we can just short circuit it altogether and bail when attempting to call these during builds.

### Why?

VersionedContentMap does a lot of hashing to account for files changing during development. In production, this is pointless as we assume a static snapshot of the repo.

### How?

Replace it with an Option and bail in the case of not being configured.