make sure all the crates in turbopack are considered when caching (#68593)
### What?

We use turbo to build the native portion of nextjs / turbopack which relies on turborepo for caching, with manually specified inputs. When moving the project we missed adding the new crates folder to the `turbo.json` file, leading to erroneous cache hits.

### Why?

Erroneous cache hits = bad.

### How?

Add the new path to the turbo.json file for next-swc. I did not include the `turbopack/packages` folder, since it is not used in native builds.